requirements.txt
Scrapy
mysql-connector-python
